linter fix

+++ controller/internal/enforcer/nfqdatapath/nfq_linux.go
@@ -104,7 +104,7 @@ func (d *Datapath) processNetworkPacketsFromNFQ(p *nfqueue.NFPacket) {
 		)
 		length := uint32(len(p.Buffer))
 		buffer := p.Buffer
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 0, uint32(uint32(p.Mark)&(^(markconstants.NFQueueMask | markconstants.NFSetMarkMask))), length, uint32(p.ID), buffer)
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 0, (p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), length, uint32(p.ID), buffer)
 		if netPacket.IPProto() == packet.IPProtocolTCP {
 			d.collectTCPPacket(&debugpacketmessage{
 				Mark:    p.Mark,
@@ -142,9 +142,9 @@ func (d *Datapath) processNetworkPacketsFromNFQ(p *nfqueue.NFPacket) {
 		copyIndex += copy(buffer[copyIndex:], netPacket.GetTCPOptions())
 		copyIndex += copy(buffer[copyIndex:], netPacket.GetTCPData())
 
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), v, uint32(uint32(p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask)), uint32(copyIndex), uint32(p.ID), buffer)
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), v, (p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), uint32(copyIndex), uint32(p.ID), buffer)
 	} else {
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), v, uint32(uint32(p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask)), uint32(len(netPacket.GetBuffer(0))), uint32(p.ID), netPacket.GetBuffer(0))
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), v, (p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), uint32(len(netPacket.GetBuffer(0))), uint32(p.ID), netPacket.GetBuffer(0))
 	}
 
 	if netPacket.IPProto() == packet.IPProtocolTCP {
@@ -202,7 +202,7 @@ func (d *Datapath) processApplicationPacketsFromNFQ(p *nfqueue.NFPacket) {
 
 		length := uint32(len(p.Buffer))
 		buffer := p.Buffer
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 0, uint32(uint32(p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask)), length, uint32(p.ID), buffer)
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 0, p.Mark&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), length, uint32(p.ID), buffer)
 		if appPacket.IPProto() == packet.IPProtocolTCP {
 
 			d.collectTCPPacket(&debugpacketmessage{
@@ -233,10 +233,10 @@ func (d *Datapath) processApplicationPacketsFromNFQ(p *nfqueue.NFPacket) {
 		copyIndex += copy(buffer[copyIndex:], appPacket.GetTCPOptions())
 		copyIndex += copy(buffer[copyIndex:], appPacket.GetTCPData())
 
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 1, uint32(uint32(p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask)), uint32(copyIndex), uint32(p.ID), buffer)
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 1, p.Mark&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), uint32(copyIndex), uint32(p.ID), buffer)
 
 	} else {
-		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 1, uint32(uint32(p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask)), uint32(len(appPacket.GetBuffer(0))), uint32(p.ID), appPacket.GetBuffer(0))
+		p.QueueHandle.SetVerdict2(uint32(p.QueueHandle.QueueNum), 1, (p.Mark)&^(markconstants.NFQueueMask|markconstants.NFSetMarkMask), uint32(len(appPacket.GetBuffer(0))), uint32(p.ID), appPacket.GetBuffer(0))
 	}
 
 	if appPacket.IPProto() == packet.IPProtocolTCP {
